Reorganized file locations, deleted BasicTeleOp.kt, and moved Incremented systems to backlog package
