+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/backlog/IncSemiAutoArm.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.subassemblies.miles.arm
+package org.firstinspires.ftc.teamcode.backlog
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode
 import com.qualcomm.robotcore.hardware.DcMotor
